graphics: Move cache clearing into ChimaraGlk
This would otherwise cause a crash on exit because clear_image_cache()
was called from the UI thread, but tried to use the thread-private data
from the Glk thread. I'm not sure how this ever worked!
